==> 1.1.procesado_datos_nuevos_entrez.ipynb <==
import pandas as pd
import os

datos={}
for i in os.listdir('datos/entrez/'):
    datos[i.split("_")[1]]=i
    # print(i.split("_")[1])

df=[]
for k,v in datos.items():
    if 'completo' not in v:
        temp=pd.read_csv(f"datos/entrez/{v}")
        temp['genus']=[k]*len(temp)
        df.append(temp)

df=pd.concat(df, ignore_index=True)

len(df)

base=set('ACGT')
drop_idx=[]
for i,j in enumerate(df['sequence']):
    if set(j)!=base:
        drop_idx.append(i)
print(len(drop_idx))
df=df.drop(drop_idx)
df=df.reset_index()

len(df)

df['genus']=df['genus'].str.capitalize()

df[[
    'index', 
    'sequence_id', 
    'sequence', 
    'sequence_length', 
    'gc_content',
    'domain', 
    'phylum', 
    'class', 
    'order', 
    'family', 
    'genus', 
    'species'
    ]].to_csv("datos/clean_entrez.csv", index=False)
